requirements.txt
deap==1.3.1
higher==0.2.1
matplotlib==3.3.3
numpy==1.19.3
PyYAML==6.0
scikit_learn==1.0.2
setuptools==51.0.0
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.54.1
pyvista==0.33.3
